pyproject.toml
[project]
name = "treecko-bot-telegram"
version = "0.1.0"
description = "Personal finance Telegram bot that processes MercadoPago transactions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=21.6",
    "pdfplumber>=0.11.4",
    "gspread>=6.1.2",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.1",
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.21.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
treecko-bot = "treecko_bot.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = []
